feat(payment): STRIPE-52 move all stripe files into stripe folder
