+++ packages/core/src/app/shipping/stripeUPE/StripeShipping.spec.tsx
@@ -1,6 +1,6 @@
 import { mount } from 'enzyme';
 import React, { FunctionComponent } from 'react';
-import CheckoutStepType from '../checkout/CheckoutStepType';
+import CheckoutStepType from '../../checkout/CheckoutStepType';
 import StripeShipping, { StripeShippingProps } from './StripeShipping';
 import StripeShippingForm from './StripeShippingForm';
